requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
sqlparse==0.4.2
python-decouple==3.6
psycopg2-binary==2.9.3
psycopg2==2.9.3
djangorestframework==3.13.1
pytz==2022.1
Pillow==9.2.0
black==22.6.0
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
tomli==2.0.1
typing-extensions==4.3.0
PyYAML==6.0
drf-spectacular==0.23.1
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.9.1
pkgutil-resolve-name==1.3.10
pyrsistent==0.18.1
uritemplate==4.1.1
zipp==3.8.1
django-debug-toolbar==3.6.0

